# docs/external-api/openapi.yaml
openapi: 3.1.0
info:
  title: Kairos CRM — External API
  version: 1.0.0
  description: |
    API para integração com apps externos (App Oficina, App Imobiliária, App Igreja, etc).

    ## Autenticação

    1. Cliente envia `POST /api/v1/external/auth/token` com:
       - Headers: `X-Kairos-App-Id`, `X-Kairos-Key-Id`, `X-Kairos-Timestamp`, `X-Kairos-Signature`
       - Body: `{ "keyId": "<prefix>", "timestamp": <unix> }`
       - Signature: `t=<unix>,v1=<hex-hmac-sha256>` onde `v1 = HMAC-SHA256(secret, "${timestamp}.${rawBody}")`

    2. Servidor responde com `{ authToken, expiresIn, scopes, tenantId }`
    3. Cliente usa `Authorization: Bearer <authToken>` em chamadas subsequentes

    ## Scopes disponíveis

    - `crm.contact.read/write`, `crm.lead.read/write`, `crm.product.read/write`
    - `crm.service.read/write`, `crm.appointment.read/write`, `crm.followup.read/write`
    - `conversation.read/write`, `ai.analyze`, `ai.chat`
    - `webhooks.read/write`, `quote.read/write`

    ## Multi-tenant

    Cada `Application` é vinculada a UM `Tenant` no momento da criação. O `tenantId`
    é derivado da credencial (NUNCA do body). Toda query é automaticamente
    escopada ao tenant do app.

    ## Webhooks outbound

    Apps podem receber eventos do CRM. Configurar via `POST /api/v1/external/webhooks`
    com `url` + `events` + `secret`. Eventos são entregues com HMAC-SHA256
    assinado em `X-Kairos-Signature: t=<ts>,v1=<hex>`.
  contact:
    name: Kairós Tecnologia
...
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Token opaco de 256 bits retornado por `POST /auth/token`.
        Header: `Authorization: Bearer <authToken>`
        TTL: 15min (default), renovável via novo `POST /auth/token`
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, TENANT_MISMATCH, RATE_LIMITED, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: object
              nullable: true

    # ===========================================
    # Auth
    # ===========================================
    AuthTokenRequest:
      type: object
      required: [keyId, timestamp]
      properties:
        keyId:
          type: string
          minLength: 8
          maxLength: 32
          description: "Prefixo da ApiKey (ex: kairos_abc12345)"
          example: "kairos_ofc01a3f1"
        timestamp:
          type: integer
          format: int64
          description: "Unix timestamp em segundos. Tolerância: ±300s"
          example: 1737043200
    AuthTokenResponse:
      type: object
      required: [authToken, expiresIn, scopes, tenantId]
      properties:
        authToken:
          type: string
          description: "Token opaco de ~43 chars (base64url). Use em Authorization: Bearer"
          example: "abc123_def456-ghi789_jkl012-mno345-pqr678"
        expiresIn:
          type: integer
          description: "TTL em segundos"
          example: 900
        expiresAt:
          type: string
          format: date-time
        scopes:
          type: array
          items: { type: string }
          example: ["crm.contact.write", "crm.lead.write", "ai.analyze"]
        tenantId:
          type: string
          format: uuid
...
    WebhookEndpoint:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string, enum: [lead.created, lead.updated, lead.stage_changed, lead.won, lead.lost, contact.created, contact.updated, message.received, message.sent, appointment.created, appointment.confirmed, appointment.cancelled, followup.scheduled, followup.sent, followup.failed, conversation.created, conversation.closed] }
        active: { type: boolean }
        description: { type: string, nullable: true }
        secretLastFour: { type: string }
        createdAt: { type: string, format: date-time }
    WebhookEndpointCreateRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          example: "https://meu-app.com/webhook"
        events:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: string
            enum: [lead.created, lead.updated, lead.stage_changed, lead.won, lead.lost, contact.created, contact.updated, message.received, message.sent, appointment.created, appointment.confirmed, appointment.cancelled, followup.scheduled, followup.sent, followup.failed, conversation.created, conversation.closed]
        description:
          type: string
          maxLength: 200
    WebhookEndpointCreateResponse:
      allOf:
        - $ref: '#/components/schemas/WebhookEndpoint'
        - type: object
          properties:
            secret:
              type: string
              description: "🔐 SECRET RETORNADO APENAS 1 VEZ. Guarde com segurança."
              example: "whsec_a3f1b2c4d5e6f7g8..."
...
paths:
...
  /api/v1/external/auth/token:
    post:
      tags: [Auth]
      summary: Trocar API Key + HMAC por authToken opaco (TTL 15min)
      description: |
        **Fluxo de autenticação**:
        1. App computa `signature = HMAC-SHA256(secret, "${timestamp}.${body}")`
        2. App envia POST com headers de assinatura + body JSON `{keyId, timestamp}`
        3. Servidor valida HMAC, lookup ApiKey, decrypt secret, verify signature
        4. Se válido: gera authToken, persiste hash no DB, retorna
        5. App usa `Authorization: Bearer <authToken>` por 15min
      security: []
      parameters:
        - name: X-Kairos-App-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
        - name: X-Kairos-Key-Id
          in: header
          required: true
          schema: { type: string, example: "kairos_ofc01a3f1" }
        - name: X-Kairos-Timestamp
          in: header
          required: true
          schema: { type: string, example: "1737043200" }
        - name: X-Kairos-Signature
          in: header
          required: true
          schema: { type: string, example: "t=1737043200,v1=abc123..." }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthTokenRequest'
      responses:
        '200':
          description: Token emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Assinatura inválida, app inativa, etc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/external/webhooks:
    get:
      tags: [Webhooks]
      summary: Lista endpoints do app
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/WebhookEndpoint' }
      security:
        - BearerAuth: ['webhooks.read']

    post:
      tags: [Webhooks]
      summary: Criar endpoint (retorna secret 1x)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointCreateRequest'
      responses:
        '201':
          description: Endpoint criado (secret retornado!)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpointCreateResponse'
      security:
        - BearerAuth: ['webhooks.write']
  /api/v1/external/webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Webhooks]
      summary: Detalha 1 endpoint
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookEndpoint' }
      security:
        - BearerAuth: ['webhooks.read']

    patch:
      tags: [Webhooks]
      summary: Atualizar URL/eventos/active
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookEndpoint' }
      security:
        - BearerAuth: ['webhooks.write']

    delete:
      tags: [Webhooks]
      summary: Desativar endpoint (soft)
      responses:
        '204':
          description: Desativado
      security:
        - BearerAuth: ['webhooks.write']
...